feat: bundle for xspec
The strategy is to compile a static library with all the gizmos we need,
and then used `initpackage` in XSPEC to get the sample Makefile and
wrapper materials. The Makefile is then modified to also link against
are static library, and we're done.

To use, a user copies the static library into the xspec directory, then
within xspec runs

        XSPEC> hmake
        XSPEC> lmod klineprofile .

and profits.

Tweaked the setup to include `Eline` parameter. Reduced the number of
radial bins and the refinement for a bit more speed (XSPEC is passing
~5000 bins often, so refinement can be more granular).